export/requirements-render.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
psutil>=5.9.5
python-dateutil>=2.8.2
wavelink>=2.6.4
pynacl>=1.5.0
stripe>=7.0.0
requests>=2.31.0
pillow>=10.0.0
jishaku>=2.5.1
python-jose>=3.3.0
attrs>=23.1.0
asyncio>=3.4.3
beautifulsoup4>=4.12.2
dblpy>=0.4.0
dnspython>=2.4.2
ffmpeg-python>=0.2.0
humanize>=4.8.0
pyfiglet>=1.0.2
pymongo>=4.5.0
tabulate>=0.9.0
youtube-dl>=2021.12.17